Update fire alarm
In a real life situation if there is a fire and smoke turning of the fire alarm would only lead to it starting up again until the fire is properly distinguished. So after many different approaches this is the one I've decided to used. There will be no turn fire alarm on, that is useless in an everyday situation. So the fire simulation switch starts the fire, which starts the alarm, is the fire is slacked the alarm will stop after a delay. :-)