remove turn fire alarm on
in real life scenario a person would not need to turn on fire alarm. when fire is simulated by the switch it will be turned on. when the switch is switched off the alarm will be switched off after short delay. if alarm switched of by mobile phone user and fire still burning alarm starts again.